+++ CHANGELOG.md
@@ -1,11 +1,18 @@
+## 0.2.0
+
+* **BREAKING**: Loaders must implement
+  [`AssetLoader`](https://pub.dev/documentation/easy_localization/latest/easy_localization/AssetLoader-class.html)
+  from easy_localization.
+* Deleted `addAllRecursive` from an extension on `Map` because easy_localization merged it.
+
 ## 0.1.2
 
-- Fix readme formatting.
+* Fix readme formatting.
 
 ## 0.1.1
 
-- Added an example.
+* Added an example.
 
 ## 0.1.0
 
-Initial release.
+* Initial release.